+++ spb/backends/matplotlib.py
@@ -398,9 +398,15 @@ class MatplotlibBackend(Plot):
                     # quivers. Setting zorder appears to fix the problem.
                     ckw["zorder"] = 0
                 kw = merge({}, ckw, s.rendering_kw)
-                c = self.ax.contourf(x, y, z, **kw)
-                self._add_colorbar(c, s.get_label(self._use_latex), s.use_cm, True)
-                self._add_handle(i, c, kw, self._fig.axes[-1])
+                func = self.ax.contourf if s.is_filled else self.ax.contour
+                c = func(x, y, z, **kw)
+                clabel = None
+                if s.is_filled:
+                    self._add_colorbar(c, s.get_label(self._use_latex),
+                        s.use_cm, True)
+                else:
+                    clabel = self.ax.clabel(c)
+                self._add_handle(i, c, kw, self._fig.axes[-1], clabel)
 
             elif s.is_3Dline:
                 x, y, z, param = s.get_data()
@@ -905,11 +911,19 @@ class MatplotlibBackend(Plot):
 
                 elif s.is_contour and (not s.is_complex):
                     x, y, z = self.series[i].get_data()
-                    kw, cax = self._handles[i][1:]
+                    kw, cax, clabels = self._handles[i][1:]
                     for c in self._handles[i][0].collections:
                         c.remove()
-                    self._handles[i][0] = self.ax.contourf(x, y, z, **kw)
-                    self._update_colorbar(cax, kw["cmap"], s.get_label(self._use_latex), param=z)
+                    if not s.is_filled:
+                        for cl in clabels:
+                            cl.remove()
+                    func = self.ax.contourf if s.is_filled else self.ax.contour
+                    self._handles[i][0] = func(x, y, z, **kw)
+                    if s.is_filled:
+                        self._update_colorbar(cax, kw["cmap"], s.get_label(self._use_latex), param=z)
+                    else:
+                        clabels = self.ax.clabel(self._handles[i][0])
+                        self._handles[i][-1] = clabels
                     xlims.append((np.amin(x), np.amax(x)))
                     ylims.append((np.amin(y), np.amax(y)))